requirements.txt
asgiref==3.6.0
Django==4.1.5
django-countries==7.5
Pillow==9.4.0
psycopg2-binary==2.9.5
python-decouple==3.6
sqlparse==0.4.3
typing_extensions==4.4.0
